+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -3,9 +3,8 @@ package techcourse.myblog.web;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import techcourse.myblog.domain.Article;
-import techcourse.myblog.domain.ArticleDto;
-import techcourse.myblog.domain.ArticleService;
+import techcourse.myblog.dto.ArticleDto;
+import techcourse.myblog.service.ArticleService;
 
 @Controller
 public class ArticleController {
